+++ meal/src/main/java/pl/auctane/meal/repositories/ProductMealRepository.java
@@ -6,6 +6,6 @@ import pl.auctane.meal.entities.ProductMeal;
 
 import java.util.List;
 
-public interface ProductMealRepository extends JpaRepository<ProductMeal, Integer> {
+public interface ProductMealRepository extends JpaRepository<ProductMeal, Long> {
     List<ProductMeal> getAllByProductId(Product product);
 }
